Added Buffer Overflow lesson, added and updated solutions, fixed errors in lessons preventing them from being completed correctly
git-svn-id: http://webgoat.googlecode.com/svn/trunk/webgoat@436 4033779f-a91e-0410-96ef-6bf7bf53c507

+++ src/main/java/org/owasp/webgoat/lessons/RoleBasedAccessControl/DeleteProfile.java
@@ -62,8 +62,9 @@ public class DeleteProfile extends DefaultLessonAction
 
 		if (isAuthenticated(s))
 		{
+			if (userId != employeeId) {
 			deleteEmployeeProfile(s, userId, employeeId);
-
+			}
 			try
 			{
 				chainedAction.handleRequest(s);
